+++ libraries/search_tools.rb
@@ -35,6 +35,15 @@ class RealityForge #nodoc
 
         mash
       end
+
+      # Workaround for versions of chef with a bug in the data bag search
+      def databag_search_prefix
+        if Chef::VERSION <= '11.16.2'
+          'raw_data_'
+        else
+          ''
+        end
+      end
     end
   end
 end