+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.2.10:
 
+* Support an array of expected types in method RealityForge::AttributeTools.get_attribute
+
 ## 0.2.8:
 
 * Improve attribute handling so that the value can be false.